==> notebooks/02-process-raw-data.ipynb <==
# Global Imports
import glob
import os
import warnings
from pathlib import Path

import geopandas as gpd
import matplotlib.pyplot as plt

# import xagg as xa
import numpy as np
import pandas as pd
import rioxarray
import xarray as xr
from shapely import wkt
from shapely.geometry import mapping

warnings.filterwarnings("ignore", message=".*initial implementation of Parquet.*")

region_file = "../data/raw/region/region.geojson"

gdf = gpd.read_file(region_file, crs="EPSG:4326")

# Clip CHIRPS to region

chirps_dir = "../data/raw/chirps/*.tif"

chirps_list = sorted(glob.glob(chirps_dir))

chirps_time = xr.Variable(
    "time", [pd.to_datetime(f"{Path(_f).name[:-4]}") for _f in chirps_list]
)

chirps_ds = xr.concat(
    [
        xr.open_rasterio(f).squeeze("band").drop("band").to_dataset(name="precip")
        for f in chirps_list
    ],
    dim=chirps_time,
)

monthly_chirps = (
    chirps_ds.resample(time="1MS").mean(dim="time").rename({"y": "lat", "x": "lon"})
)

monthly_chirps.to_netcdf("../data/processed/chirps.nc")

# Convert Terra Data to netcdf

terra_dir = "../data/raw/terra_climate/*.tif"

terra_list = sorted(glob.glob(terra_dir))

Path(terra_list[0]).name[:-4]

terra_time = xr.Variable(
    "time", [pd.to_datetime(f"{Path(_f).name[:-4]}01") for _f in terra_list]
)

terra_climate_bands = ["aet", "def", "pdsi", "pet", "pr", "srad", "ro", "soil", "swe"]

terra_da = xr.concat(
    [xr.open_rasterio(f).rename({"y": "lat", "x": "lon"}) for f in terra_list],
    dim=terra_time,
)

terra_da["band"] = terra_climate_bands

terra_ds = terra_da.to_dataset(dim="band")

fig, ax = plt.subplots(1, 2)
terra_ds.pdsi.isel(time=100).plot(ax=ax[1])
gdf.plot(ax=ax[1], edgecolor="red", color="none")
chirps_ds.precip.isel(time=100).plot(ax=ax[0])
gdf.plot(ax=ax[0], edgecolor="red", color="none")

terra_ds.to_netcdf("../data/processed/terra.nc")

terra_ds.aet.isel(time=10).plot()

chirps_ds.precip.isel(time=10).plot()

# Clip GRACE
# Loading GRACE dataset
grace = xr.open_dataset(
    "../data/raw/mascon/GRCTellus.JPL.200204_202206.GLO.RL06M.MSCNv02CRI.nc"
)["lwe_thickness"].to_dataset()

# Calculate GRACE TWS Anomalies
# Clipping the grace ds to mississippi shapefile
grace = grace.assign({"lon": (((grace.lon + 180) % 360) - 180)}).sortby("lon")
grace["lwe_thickness"] = grace["lwe_thickness"].rio.write_crs("epsg:4326")
grace.rio.set_spatial_dims(x_dim="lon", y_dim="lat", inplace=True)

grace_clipped = grace.rio.clip(
    gdf.geometry.apply(mapping), gdf.crs, all_touched=True, drop=True
)

final_grace = grace_clipped.resample(time="1MS").mean(dim="time")

final_grace.to_netcdf("../data/processed/grace.nc")
