==> notebooks/04-deep-learning-model-execution.ipynb <==
import glob

import numpy as np
import pandas as pd
import tensorflow as tf
import tensorflow_addons as tfa
import xarray as xr
from keras.layers import Dense
from keras.models import Sequential
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

model_df = pd.read_csv("../data/processed/model_inputs.csv")

list(model_df)

input_df = model_df[
    [
        "lat_2",
        "lon_2",
        "time",
        "aet",
        "def",
        "pdsi",
        "pet",
        "pr",
        "srad",
        "ro",
        "soil",
        "swe",
        "precip",
        "lwe_thickness",
    ]
]
input_df["precip"] = input_df["precip"].fillna(0)

input_df = input_df.dropna().reset_index(drop=True)

input_df

X_df = input_df.iloc[:, 0:13]

y_df = input_df.iloc[:, -1]

time_dict = dict(zip(X_df["time"].unique().tolist(), range(0, X_df.time.nunique())))

X_df["time"] = X_df.time.map(time_dict)

X_train, X_test, y_train, y_test = train_test_split(
    X_df.to_numpy(), y_df.to_numpy(), test_size=0.2
)

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)

X_test_scaled = scaler.transform(X_test)

n_features = X_train.shape[1]

# Setting up error metrics
rmse = tf.keras.metrics.RootMeanSquaredError(name="root_mean_squared_error", dtype=None)

model = Sequential()
model.add(
    Dense(
        48, activation="sigmoid", kernel_initializer="normal", input_shape=(n_features,)
    )
)
model.add(Dense(40, activation="relu", kernel_initializer="he_normal"))
model.add(Dense(36, activation="relu", kernel_initializer="he_normal"))
model.add(Dense(24, activation="relu", kernel_initializer="he_normal"))
model.add(Dense(18, activation="relu", kernel_initializer="he_normal"))
model.add(Dense(12, activation="relu", kernel_initializer="he_normal"))
model.add(Dense(1))
# compile the model
model.compile(
    optimizer=tf.keras.optimizers.Adam(0.01), loss="mse", metrics=["mae", rmse]
)

model.fit(X_train_scaled, y_train, epochs=125, batch_size=len(X_df[X_df.time == 0]))

error = model.evaluate(X_train_scaled, y_train, verbose=0)

error

X_df[X_df.time == 1]

y_pred = model.predict(X_test_scaled)

from sklearn import metrics

print("Mean Absolute Error:", metrics.mean_absolute_error(y_test, y_pred))
print("Mean Squared Error:", metrics.mean_squared_error(y_test, y_pred))
print("Root Mean Squared Error:", np.sqrt(metrics.mean_squared_error(y_test, y_pred)))

metrics.r2_score(y_test, y_pred)

model.save("../models/dl_model")

model = tf.keras.models.load_model("../models/dl_model/")

def predict_ts(ts_file):
    ts_df = pd.read_parquet(ts_file)
    ts_df["time"] = ts_df.time.dt.strftime("%Y-%m-%d").map(time_dict)
    ts_test_x = ts_df.to_numpy()
    ts_step_scaled = scaler.transform(ts_test_x)
    ts_pred = model.predict(ts_step_scaled)
    output_df = pd.DataFrame(ts_pred, columns=["lwe_thickness"])
    final_ts_df = pd.concat([ts_df[["lat", "lon", "time"]], output_df], axis=1)
    final_ts_df["time"] = final_ts_df["time"].map(
        dict(zip(time_dict.values(), time_dict.keys()))
    )
    final_ts_df["time"] = pd.to_datetime(final_ts_df.time)
    ds = final_ts_df.groupby(["lat", "lon", "time"]).mean().to_xarray()
    return ds

xr.concat(
    [
        predict_ts(_file)
        for _file in sorted(glob.glob("../data/processed/target_grid/*"))
    ],
    dim="time",
).to_netcdf("../data/processed/dl_output.nc")

predicted_grace = xr.open_dataset("../data/processed/dl_output.nc")

actual_grace = xr.open_dataset("../data/processed/grace.nc")

actual_grace

predicted_grace

model_df[model_df.time == "2002-04-01"]

input_df[input_df.time == "2002-04-01"]

import matplotlib.pyplot as plt

fig, ax = plt.subplots(1, 2, figsize=(15, 10))
predicted_grace.lwe_thickness.sel(time="2008-01-01").plot(ax=ax[1])
actual_grace.lwe_thickness.sel(time="2008-01-01").plot(ax=ax[0])
# plt.savefig("../reports/figures/actual_vs_dl.png")

import matplotlib.pyplot as plt

fig, ax = plt.subplots(1, 2, figsize=(15, 10))
predicted_grace.lwe_thickness.sel(time="2008-01-01").plot(ax=ax[1])
actual_grace.lwe_thickness.sel(time="2008-01-01").plot(ax=ax[0])
# plt.savefig("../reports/figures/actual_vs_dl.png")

import matplotlib.pyplot as plt

fig, ax = plt.subplots(1, 1, figsize=(25, 10))
predicted_grace.lwe_thickness.mean(["lat", "lon"]).plot(ax=ax, label="Predicted")
actual_grace.lwe_thickness.mean(["lat", "lon"]).plot(ax=ax, label="Original")
plt.legend()
# plt.savefig("mean.png")
